requirements.txt
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pycryptodome==3.20.0
pytest==8.2.2
